feat(perf): Optimize EditorScreen rendering with memoized ActionItem
Extracted ActionItem component and refactored EditorScreen to use stable callbacks, preventing unnecessary re-renders of the entire action list when editing a single action. This significantly improves performance for tasks with many actions.
